+++ Project/drivers/elevator/elevator.go
@@ -81,6 +81,6 @@ func SetElevatorBehaviour(elev *Elevator, behavior ElevBehavior) {
 	elev.Behavior = behavior
 }
 
-func SetElevatorOpenDuration(elev *Elevator, time_s float) {
+func SetElevatorOpenDuration(elev *Elevator, time_s float64) {
 	elev.OpenDuration = time_s
 }
